Add files via upload
This project is all about analyzing a Delhivery trucks trip in entire country. I used Excel to go through a lot of data and understand different aspects of the trip in logistics. Delhivery, the biggest logistics company in India, inspired me to showcase my data skills in a practical way. Whether it's dealing with business (B2B), customers (B2C), or personal connections (C2C).
The project demonstrates how Excel can be a powerful tool for understanding and sharing insights. It's a hands-on example of how I can apply data analysis to real-world situations.